+++ tests/rules/test_result.py
@@ -1,4 +1,5 @@
-from arche.rules.result import Message
+from arche.rules.result import Level, Message
+from conftest import create_result
 import pandas as pd
 import pytest
 
@@ -36,3 +37,19 @@ def test_message_eq(source, target):
 )
 def test_message_not_eq(source, target):
     assert Message(*source) != Message(*target)
+
+
+@pytest.mark.parametrize(
+    "message, expected_details",
+    [
+        (
+            {Level.INFO: [("summary", "very detailed message")]},
+            "\nrule name here:\n\tsummary\nvery detailed message\n",
+        ),
+        ({Level.INFO: [("summary",)]}, "\nrule name here:\n\tsummary\n"),
+    ],
+)
+def test_show(capsys, message, expected_details):
+    r = create_result("rule name here", message)
+    r.show()
+    assert capsys.readouterr().out == expected_details